+++ src/main/java/de/chojo/repbot/commands/bot/BotAdmin.java
@@ -54,8 +54,8 @@ public class BotAdmin extends SlashCommand {
                    .group(Group.of("log", "log access")
                                .subCommand(SubCommand.of("analyzer", "Analyzer log")
                                                      .handler(new Analyzer(guilds))
-                                                     .argument(Argument.text("guild_id", "Guild id").asRequired())
-                                                     .argument(Argument.text("message_id", "Id of message").asRequired())))
+                                                     .argument(Argument.text("guildid", "Guild id").asRequired())
+                                                     .argument(Argument.text("messageid", "Id of message").asRequired())))
                    .subCommand(SubCommand.of("shared_guilds", "Shared guilds with a user")
                                          .handler(new SharedGuilds(configuration))
                                          .argument(Argument.text("user_id", "user id"))
